requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.0
asttokens==3.0.1
attrs==25.4.0
beautifulsoup4==4.14.3
brotli==1.2.0
certifi==2026.1.4
chardet==3.0.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
comm==0.2.3
dataclasses-json==0.6.7
ddgs==9.10.0
debugpy==1.8.19
decorator==5.2.1
deep-translator==1.11.4
distro==1.9.0
Django==6.0.1
django-modeltranslation==0.19.19
django-rosetta==0.10.3
djangorestframework==3.16.1
executing==2.2.1
fake-useragent==2.2.0
filetype==1.2.0
frozenlist==1.8.0
google-auth==2.47.0
google-genai==1.59.0
googletrans==4.0.0rc1
greenlet==3.3.0
h11==0.16.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==1.0.9
httpx==0.27.2
httpx-sse==0.4.3
hyperframe==5.2.0
idna==2.10
ipykernel==7.1.0
ipython==9.9.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.8.0
jupyter_core==5.9.1
langchain==1.2.4
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.7
langchain-google-genai==4.2.0
langchain-text-splitters==1.1.0
langgraph==1.0.6
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.6
langgraph-sdk==0.3.3
langsmith==0.6.4
legacy-cgi==2.6.4
lxml==6.0.2
marshmallow==3.26.2
matplotlib-inline==0.2.1
multidict==6.7.0
mypy_extensions==1.1.0
nest-asyncio==1.6.0
numpy==2.4.1
orjson==3.11.5
ormsgpack==1.12.1
packaging==25.0
parso==0.8.5
platformdirs==4.5.1
polib==1.2.0
primp==0.15.0
prompt_toolkit==3.0.52
propcache==0.4.1
psutil==7.2.1
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
pyzmq==27.1.0
requests==2.32.5
requests-toolbelt==1.0.0
rfc3986==1.5.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.8.1
SQLAlchemy==2.0.45
sqlparse==0.5.5
stack-data==0.6.3
tenacity==9.1.2
tornado==6.5.4
traitlets==5.14.3
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uuid_utils==0.13.0
wcwidth==0.2.14
websockets==15.0.1
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
